+++ test/output.jl
@@ -69,6 +69,53 @@ end
         @test isnan(branches["3"]["qf"])
         @test isnan(branches["3"]["qt"])
     end
+
+    @testset "24-bus rts case ac pf" begin
+        result = run_pf("../test/data/matpower/case24.m", ACPPowerModel, ipopt_solver; setting = Dict("output" => Dict("branch_flows" => true)))
+
+        @test haskey(result, "optimizer") == true
+        @test haskey(result, "termination_status") == true
+        @test haskey(result, "primal_status") == true
+        @test haskey(result, "dual_status") == true
+        @test haskey(result, "objective") == true
+        @test haskey(result, "objective_lb") == true
+        @test haskey(result, "solve_time") == true
+        @test haskey(result, "machine") == true
+        @test haskey(result, "data") == true
+        @test haskey(result, "solution") == true
+        @test haskey(result["solution"], "branch") == true
+
+        @test length(result["solution"]["bus"]) == 24
+        @test length(result["solution"]["gen"]) == 33
+        @test length(result["solution"]["branch"]) == 38
+
+        branches = result["solution"]["branch"]
+
+        @test isapprox(branches["2"]["pf"],  0.2001; atol = 1e-3)
+        @test isapprox(branches["2"]["pt"], -0.1980; atol = 1e-3)
+        @test isapprox(branches["2"]["qf"],  0.0055; atol = 1e-3)
+        @test isapprox(branches["2"]["qt"], -0.0571; atol = 1e-3)
+    end
+
+    # A DCPPowerModel test is important because it does have variables for the reverse side of the branchs
+    @testset "3-bus case dc pf" begin
+        result = run_pf("../test/data/matpower/case3.m", DCPPowerModel, ipopt_solver; setting = Dict("output" => Dict("branch_flows" => true)))
+
+        @test haskey(result, "solution") == true
+        @test haskey(result["solution"], "branch") == true
+
+        @test length(result["solution"]["bus"]) == 3
+        @test length(result["solution"]["gen"]) == 3
+        @test length(result["solution"]["branch"]) == 3
+        @test length(result["solution"]["dcline"]) == 1
+
+        branches = result["solution"]["branch"]
+
+        @test isapprox(branches["3"]["pf"], -0.101419; atol = 1e-3)
+        @test isapprox(branches["3"]["pt"],  0.101419; atol = 1e-3)
+        @test isnan(branches["3"]["qf"])
+        @test isnan(branches["3"]["qt"])
+    end
 end
 
 
@@ -143,7 +190,7 @@ end
 end
 
 
-# recomended by @lroald
+# recommended by @lroald
 @testset "test solution feedback" begin
 
     function solution_feedback(case, ac_opf_obj)